+++ Rainbow/shared/utilities.hpp
@@ -4,8 +4,8 @@ namespace rainbow {
 
 	using int32 = int;
 	
-	using uint32 = unsigned long long;
-	using uint64 = unsigned;
+	using uint32 = unsigned;
+	using uint64 = unsigned long long;
 
 	using real = float;
